fix float64 to float32

+++ internal/models/balance.go
@@ -5,15 +5,15 @@ import "time"
 type Balance struct {
 	ID        uint64    `gorm:"primary_key"`
 	UserID    uint64    `gorm:"not null;unique"`
-	Current   float64   `gorm:"default:0"`
-	Withdraw  float64   `gorm:"default:0"`
+	Current   float32   `gorm:"default:0"`
+	Withdraw  float32   `gorm:"default:0"`
 	ProcessAt time.Time `gorm:"autoCreateTime"`
 }
 
 type BalanceJSON struct {
 	ID        uint64    `gorm:"primary_key" json:"-"`
 	UserID    uint64    `gorm:"not null" json:"-"`
-	Current   float64   `gorm:"default:null" json:"current"`
-	Withdraw  float64   `gorm:"default:null" json:"withdraw"`
+	Current   float32   `gorm:"default:null" json:"current"`
+	Withdraw  float32   `gorm:"default:null" json:"withdraw"`
 	ProcessAt time.Time `gorm:"autoCreateTime" json:"-"`
 }
